requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
awscli==1.29.40
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
boto3==1.28.40
botocore==1.31.40
cachetools==5.3.1
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
click==7.1.2
colorama==0.4.4
contourpy==1.1.0
cycler==0.11.0
dnspython==2.4.1
docker==6.1.3
docutils==0.16
filelock==3.12.3
Flask==1.1.2
Flask-Cors==3.0.8
fonttools==4.42.1
frozenlist==1.4.0
future==0.18.2
gitdb==4.0.10
GitPython==3.1.34
idna==2.9
importlib-metadata==6.8.0
importlib-resources==6.0.1
itsdangerous==1.1.0
jaraco.classes==3.3.0
Jinja2==2.11.2
jmespath==1.0.1
joblib==1.3.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keyring==8.7
keyrings.alt==5.0.0
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==1.1.1
matplotlib==3.7.2
mdurl==0.1.2
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
nltk==3.8.1
numpy==1.24.4
oauthlib==3.1.0
openai==0.27.8
opencv-python==4.8.0.76
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
protobuf==4.24.2
psutil==5.9.5
py-cpuinfo==9.0.0
pyarrow==13.0.0
pyasn1==0.5.0
pydeck==0.8.0
pyee==8.2.2
Pygments==2.16.1
pymongo==4.4.1
Pympler==1.0.1
pyparsing==3.0.9
pyppeteer==1.0.2
python-dateutil==2.8.2
python-twitter==3.5
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
referencing==0.30.2
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.0
rich==13.5.2
rpds-py==0.10.2
rsa==4.7.2
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
s3transfer==0.6.2
scipy==1.10.1
seaborn==0.12.2
six==1.14.0
smmap==5.0.0
soupsieve==2.4.1
streamlit==1.26.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
ultralytics==8.0.170
urllib3==1.26.16
validators==0.22.0
voluptuous==0.13.1
watchdog==3.0.0
websocket-client==1.6.2
websockets==10.4
Werkzeug==1.0.1
yarl==1.9.2
yolo==0.3.1
zipp==3.16.2
